[skip-ci] WIP: extract only communication
