+++ save-orchestrator/src/main/kotlin/com/saveourtool/save/orchestrator/service/AgentService.kt
@@ -1,24 +1,44 @@
 package com.saveourtool.save.orchestrator.service
 
-import com.saveourtool.save.agent.*
+import com.saveourtool.save.agent.AgentState
+import com.saveourtool.save.agent.AgentState.*
+import com.saveourtool.save.agent.HeartbeatResponse
+import com.saveourtool.save.agent.NewJobResponse
+import com.saveourtool.save.agent.WaitResponse
 import com.saveourtool.save.entities.Agent
 import com.saveourtool.save.entities.AgentStatus
 import com.saveourtool.save.entities.AgentStatusDto
 import com.saveourtool.save.execution.ExecutionStatus
 import com.saveourtool.save.orchestrator.BodilessResponseEntity
-import com.saveourtool.save.orchestrator.ExecutionIdToAgentIds
+import com.saveourtool.save.orchestrator.config.ConfigProperties
+import com.saveourtool.save.orchestrator.runner.AgentRunner
+import com.saveourtool.save.test.TestBatch
+import com.saveourtool.save.test.TestDto
+import com.saveourtool.save.utils.*
+import org.slf4j.LoggerFactory
+import org.springframework.stereotype.Service
+import org.springframework.web.reactive.function.client.WebClientException
 import org.springframework.web.reactive.function.client.WebClientResponseException
 import reactor.core.publisher.Mono
 import reactor.core.scheduler.Scheduler
+import reactor.core.scheduler.Schedulers
+import reactor.kotlin.core.publisher.onErrorResume
+import java.time.Duration
+import java.time.LocalDateTime
 
 /**
  * Service for work with agents and backend
  */
-interface AgentService {
+@Service
+class AgentService(
+    private val configProperties: ConfigProperties,
+    private val agentRunner: AgentRunner,
+    private val bridgeService: BridgeService,
+) {
     /**
      * A scheduler that executes long-running background tasks
      */
-    val scheduler: Scheduler
+    val scheduler: Scheduler = Schedulers.boundedElastic().also { it.start() }
 
     /**
      * Sets new tests ids
@@ -26,7 +46,9 @@ interface AgentService {
      * @param agentId
      * @return Mono<NewJobResponse>
      */
-    fun getNewTestsIds(agentId: String): Mono<HeartbeatResponse>
+    fun getNewTestsIds(agentId: String): Mono<HeartbeatResponse> =
+            bridgeService.getNextTestBatch(agentId)
+                .flatMap { it.toHeartbeatResponse(agentId) }
 
     /**
      * Save new agents to the DB and insert their statuses. This logic is performed in two consecutive requests.
@@ -35,19 +57,25 @@ interface AgentService {
      * @return Mono with response body
      * @throws WebClientResponseException if any of the requests fails
      */
-    fun saveAgentsWithInitialStatuses(agents: List<Agent>): Mono<Void>
-
-    /**
-     * @param agentStates list of [AgentStatus]es to update in the DB
-     * @return Mono with response body
-     */
-    fun updateAgentStatuses(agentStates: List<AgentStatus>): Mono<Void>
+    fun saveAgentsWithInitialStatuses(agents: List<Agent>): Mono<Unit> = bridgeService
+        .addAgents(agents)
+        .flatMap { agentIds ->
+            bridgeService.updateAgentStatuses(agents.zip(agentIds).map { (agent, id) ->
+                AgentStatus(LocalDateTime.now(), LocalDateTime.now(), STARTING, agent.also { it.id = id })
+            })
+        }
 
     /**
      * @param agentState [AgentStatus] to update in the DB
      * @return a Mono containing bodiless entity of response or an empty Mono if request has failed
      */
-    fun updateAgentStatusesWithDto(agentState: AgentStatusDto): Mono<BodilessResponseEntity>
+    fun updateAgentStatusesWithDto(agentState: AgentStatusDto): Mono<BodilessResponseEntity> =
+            bridgeService
+                .updateAgentStatusesWithDto(agentState)
+                .onErrorResume(WebClientException::class) {
+                    log.warn("Couldn't update agent statuses because of backend failure", it)
+                    Mono.empty()
+                }
 
     /**
      * Check that no TestExecution for agent [agentId] have status READY_FOR_TESTING
@@ -55,14 +83,44 @@ interface AgentService {
      * @param agentId agent for which data is checked
      * @return true if all executions have status other than `READY_FOR_TESTING`
      */
-    fun checkSavedData(agentId: String): Mono<Boolean>
+    fun checkSavedData(agentId: String): Mono<Boolean> = bridgeService
+        .getReadyForTestingTestExecutions(agentId)
+        .map { it.isEmpty() }
 
     /**
      * This method should be called when all agents are done and execution status can be updated and cleanup can be performed
      *
      * @param agentId an ID of the agent from the execution, that will be checked.
      */
-    fun finalizeExecution(agentId: String)
+    @Suppress("TOO_LONG_FUNCTION", "AVOID_NULL_CHECKS")
+    fun finalizeExecution(agentId: String) {
+        // Get a list of agents for this execution, if their statuses indicate that the execution can be terminated.
+        // I.e., all agents must be stopped by this point in order to move further in shutdown logic.
+        getFinishedOrStoppedAgentsForSameExecution(agentId)
+            .filter { (_, finishedAgentIds) -> finishedAgentIds.isNotEmpty() }
+            .flatMap { (_, _) ->
+                // need to retry after some time, because for other agents BUSY state might have not been written completely
+                log.debug("Waiting for ${configProperties.shutdown.checksIntervalMillis} ms to repeat `getAgentsAwaitingStop` call for agentId=$agentId")
+                Mono.delay(Duration.ofMillis(configProperties.shutdown.checksIntervalMillis)).then(
+                    getFinishedOrStoppedAgentsForSameExecution(agentId)
+                )
+            }
+            .filter { (_, finishedAgentIds) -> finishedAgentIds.isNotEmpty() }
+            .flatMap { (executionId, finishedAgentIds) ->
+                log.info { "For execution id=$executionId all agents have completed their lifecycle" }
+                markExecutionBasedOnAgentStates(executionId, finishedAgentIds)
+                    .thenReturn(
+                        agentRunner.cleanup(executionId)
+                    )
+            }
+            .doOnSuccess {
+                if (it == null) {
+                    log.debug("Agents other than $agentId are still running, so won't try to stop them")
+                }
+            }
+            .subscribeOn(scheduler)
+            .subscribe()
+    }
 
     /**
      * Updates status of execution [executionId] based on statues of agents [agentIds]
@@ -74,7 +132,28 @@ interface AgentService {
     fun markExecutionBasedOnAgentStates(
         executionId: Long,
         agentIds: List<String>,
-    ): Mono<BodilessResponseEntity>
+    ): Mono<BodilessResponseEntity> {
+        // all { STOPPED_BY_ORCH || TERMINATED } -> FINISHED
+        // all { CRASHED } -> ERROR; set all test executions to CRASHED
+        return bridgeService
+            .getAgentsStatuses(executionId, agentIds)
+            .flatMap { agentStatuses ->
+                // todo: take test execution statuses into account too
+                if (agentStatuses.map { it.state }.all {
+                    it == STOPPED_BY_ORCH || it == TERMINATED
+                }) {
+                    bridgeService.updateExecutionByDto(executionId, ExecutionStatus.FINISHED)
+                } else if (agentStatuses.map { it.state }.all {
+                    it == CRASHED
+                }) {
+                    bridgeService.updateExecutionByDto(executionId, ExecutionStatus.ERROR,
+                        "All agents for this execution were crashed unexpectedly"
+                    ).then(markTestExecutionsAsFailed(agentIds, CRASHED))
+                } else {
+                    Mono.error(NotImplementedError("Updating execution (id=$executionId) status for agents with statuses $agentStatuses is not supported yet"))
+                }
+            }
+    }
 
     /**
      * Marks the execution to specified state
@@ -84,28 +163,46 @@ interface AgentService {
      * @param failReason to show to user in case of error status
      * @return a bodiless response entity
      */
-    fun updateExecution(executionId: Long, executionStatus: ExecutionStatus, failReason: String? = null): Mono<BodilessResponseEntity>
+    fun updateExecution(executionId: Long, executionStatus: ExecutionStatus, failReason: String?): Mono<BodilessResponseEntity> =
+            bridgeService.updateExecutionByDto(executionId, executionStatus, failReason)
 
     /**
      * Get list of agent ids (containerIds) for agents that have completed their jobs.
      * If we call this method, then there are no unfinished TestExecutions. So we check other agents' status.
      *
-     * We assume, that all agents will eventually have one of statuses are finished.
+     * We assume, that all agents will eventually have one of statuses [areFinishedOrStopped].
      * Situations when agent gets stuck with a different status and for whatever reason is unable to update
      * it, are not handled. Anyway, such agents should be eventually stopped by [HeartBeatInspector].
      *
      * @param agentId containerId of an agent
      * @return Mono with list of agent ids for agents that can be shut down for an executionId
      */
-    fun getFinishedOrStoppedAgentsForSameExecution(agentId: String): Mono<ExecutionIdToAgentIds>
+    @Suppress("TYPE_ALIAS")
+    fun getFinishedOrStoppedAgentsForSameExecution(agentId: String): Mono<Pair<Long, List<String>>> = bridgeService
+        .getAgentsStatusesForSameExecution(agentId)
+        .map { (executionId, agentStatuses) ->
+            log.debug("For executionId=$executionId agent statuses are $agentStatuses")
+            // with new logic, should we check only for CRASHED, STOPPED, TERMINATED?
+            executionId to if (agentStatuses.areFinishedOrStopped()) {
+                log.debug("For execution id=$executionId there are finished or stopped agents")
+                agentStatuses.map { it.containerId }
+            } else {
+                emptyList()
+            }
+        }
 
     /**
      * Checks whether all agent under one execution have completed their jobs.
      *
      * @param agentId containerId of an agent
-     * @return true if all agents match finished status
+     * @return true if all agents match [areIdleOrFinished]
      */
-    fun areAllAgentsIdleOrFinished(agentId: String): Mono<Boolean>
+    fun areAllAgentsIdleOrFinished(agentId: String): Mono<Boolean> = bridgeService
+        .getAgentsStatusesForSameExecution(agentId)
+        .map { (executionId, agentStatuses) ->
+            log.debug("For executionId=$executionId agent statuses are $agentStatuses")
+            agentStatuses.areIdleOrFinished()
+        }
 
     /**
      * Perform two operations in arbitrary order: assign `agentContainerId` agent to test executions
@@ -114,7 +211,23 @@ interface AgentService {
      * @param agentContainerId id of an agent that receives tests
      * @param newJobResponse a heartbeat response with tests
      */
-    fun updateAssignedAgent(agentContainerId: String, newJobResponse: NewJobResponse)
+    fun updateAssignedAgent(agentContainerId: String, newJobResponse: NewJobResponse) {
+        bridgeService.assignAgent(agentContainerId, newJobResponse.tests).zipWith(
+            updateAgentStatusesWithDto(
+                AgentStatusDto(LocalDateTime.now(), BUSY, agentContainerId)
+            )
+        )
+            .doOnSuccess {
+                log.trace { "Agent $agentContainerId has been set as executor for tests ${newJobResponse.tests} and its status has been set to BUSY" }
+            }
+            .subscribeOn(scheduler)
+            .subscribe()
+    }
+
+    private fun updateTestExecutionsWithAgent(agentId: String, testDtos: List<TestDto>): Mono<BodilessResponseEntity> {
+        log.trace("Attempt to update test executions for tests=$testDtos for agent $agentId")
+        return bridgeService.assignAgent(agentId, testDtos)
+    }
 
     /**
      * Mark agent's test executions as failed
@@ -123,5 +236,34 @@ interface AgentService {
      * @param status
      * @return a bodiless response entity
      */
-    fun markTestExecutionsAsFailed(agentsList: Collection<String>, status: AgentState): Mono<BodilessResponseEntity>
+    fun markTestExecutionsAsFailed(agentsList: Collection<String>, status: AgentState): Mono<BodilessResponseEntity> {
+        log.debug("Attempt to mark test executions of agents=$agentsList as failed with internal error")
+        return bridgeService.setStatusByAgentIds(agentsList, status)
+    }
+
+    private fun TestBatch.toHeartbeatResponse(agentId: String): Mono<HeartbeatResponse> =
+            if (isNotEmpty()) {
+                // fixme: do we still need suitesToArgs, since we have execFlags in save.toml?
+                Mono.fromCallable { NewJobResponse(this, constructCliCommand()) }
+            } else {
+                log.debug("Next test batch for agentId=$agentId is empty, setting it to wait")
+                Mono.just(WaitResponse)
+            }
+
+    private fun TestBatch.constructCliCommand() = joinToString(" ") { it.filePath }
+        .also {
+            log.debug("Constructed cli args for SAVE-cli: $it")
+        }
+
+    private fun Collection<AgentStatusDto>.areIdleOrFinished() = all {
+        it.state == IDLE || it.state == FINISHED || it.state == STOPPED_BY_ORCH || it.state == CRASHED || it.state == TERMINATED
+    }
+
+    private fun Collection<AgentStatusDto>.areFinishedOrStopped() = all {
+        it.state == FINISHED || it.state == STOPPED_BY_ORCH || it.state == CRASHED || it.state == TERMINATED
+    }
+
+    companion object {
+        private val log = LoggerFactory.getLogger(AgentService::class.java)
+    }
 }
